# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TaskManager DOC API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/profiles": {
            "get": {
                "tags": [
                    "Админские функции (Данные Профиля)"
                ],
                "summary": "Просмотр Профилей всех пользователей",
                "operationId": "edae69743e91e7332742bcc3b9330319",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Иван"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Иванов"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "07-05-1985"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "Noone"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/profiles/{profile}": {
            "get": {
                "tags": [
                    "Админские функции (Данные Профиля)"
                ],
                "summary": "Просмотр Профиля пользователя",
                "operationId": "c1b6d28571ae21efc886b95077816067",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "id profile",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "first_name": {
                                            "type": "string",
                                            "example": "Иван"
                                        },
                                        "last_name": {
                                            "type": "string",
                                            "example": "Иванов"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "07-05-1985"
                                        },
                                        "login": {
                                            "type": "string",
                                            "example": "Noone"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2010-10-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Админские функции (Данные Профиля)"
                ],
                "summary": "Обновление данных профиля пользователя",
                "operationId": "98934b698be07449384b973074935149",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "id profile",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "first_name": {
                                                "type": "string",
                                                "example": "Иван"
                                            },
                                            "last_name": {
                                                "type": "string",
                                                "example": "Иванов"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "07-05-1985"
                                            },
                                            "login": {
                                                "type": "string",
                                                "example": "Noone"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Иван"
                                                    },
                                                    "last_name": {
                                                        "type": "string",
                                                        "example": "Иванов"
                                                    },
                                                    "date_of_birth": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "07-05-1985"
                                                    },
                                                    "login": {
                                                        "type": "string",
                                                        "example": "Noone"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2010-10-22"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unautoruzed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}